speller.c
The code herein is designed to first load a dictionary into memory as a hash table data structure. The program then opens and reads a chosen text (as specified in the command-line argument) character by character. If a word is found which is equal to or less than 45 characters in length, then this is passed as an argument to the function "check". The latter hashes the word to return the hash table index where it is (or is not) located in memory. Since each index is a linked list of structs, with each in turn containing a word, this is followed through memory until the correct word is either found or not found using the function "strcmp". Any words which are not found are treated as "mispellings". Further functions in speller.c include "size", which calculates the total number of words in the uploaded dictionary and "unload", which frees the dictionary from memory. Throughout the running of the program, "getusage" benchmarks (times the execution of) load, check, size and unload to provide a total execution time. 

Usage: ./speller text