dictionary.c
Implements the functionality of the dictionary as declared in dictionary.h and called by speller.c. The function "load" takes the hard-coded file location of the on-disc dictionary and reads it line by line. Unless end of file is reached, a new struct is created within memory and a word stored within. This new struct is hashed (using a hash function and passing word as an argument or key) and set to point in memory to the last added struct/word (or NULL if this is the first entry at that particular index). Having a linked list at each hash table index is a way to deal with collisions (where multiple words are hashed to the same index) - a technique called "separate chaining". The function "check", takes a word from the submitted text and first converts it to all lower-case lettering. The word is then "hashed" to return the hash-table index at which it would reside in the uploaded dictionary. The linked list at this index is then traversed and strcmp employed to identify if the word is present or not. If it is not, then it is treated as a mispelling. The function "size" returns how many words are in the uploaded dictionary by indexing through the hash table and at each index traversing the linked list of structs, keeping count at each point. The function "unload" does the same, except uses "free" to unload the dictionary from memory.